+++ aepp/config.py
@@ -40,6 +40,7 @@ endpoints = {
     "dataset":"/data/foundation/dataset",
     "ingestion":"/data/foundation/import",
     "observability":"/data/infrastructure/observability/insights",
+    "destinationAuthoring" : "/data/core/activation/authoring",
     "streaming": {
         "inlet": "https://platform.adobe.io/data/core/edge",
         "collection" : "https://dcs.adobedc.net"